Added documentation to rest of the classes and methods. Delete TimedBreak class and xml activity.
