+++ app/src/main/AndroidManifest.xml
@@ -13,9 +13,6 @@
             android:name=".EndActivity"
             android:exported="false" />
         <activity
-            android:name=".TimedBreak"
-            android:exported="false" />
-        <activity
             android:name=".ExpressExerciseActivity"
             android:exported="false" />
         <activity
@@ -28,7 +25,7 @@
             android:name=".ProfileActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".SignupActivity"
             android:exported="false" />
         <activity
             android:name=".LoginActivity"
