+++ app/src/main/java/com/example/alafitness/ExpressExerciseActivity.java
@@ -39,7 +39,6 @@ public class ExpressExerciseActivity extends AppCompatActivity implements TextTo
     private ImageView exerciseImage;
     private Long progress;
     private ProgressBar timerBar;
-    private TextView startTimerView;
     private Button startPauseButton;
     private List<TimedExercise> exercises;
     private int currentExercise = 0;
@@ -76,14 +75,12 @@ public class ExpressExerciseActivity extends AppCompatActivity implements TextTo
         nextBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent2 = new Intent(ExpressExerciseActivity.this, TimedBreak.class);
-                startActivity(intent2);
+//                Intent intent2 = new Intent(ExpressExerciseActivity.this, TimedBreak.class);
+//                startActivity(intent2);
             }
         });
 
-        startTimerView = findViewById(R.id.timer_View);
         startPauseButton = findViewById(R.id.pause_Button);
-
         startPauseButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -108,7 +105,6 @@ public class ExpressExerciseActivity extends AppCompatActivity implements TextTo
             startTimer();
         }
     }
-
 
     public void startTimer() {
 
@@ -162,7 +158,6 @@ public class ExpressExerciseActivity extends AppCompatActivity implements TextTo
                 }
             }
 
-
             @Override
             public void onFinish() {
                 timerBar.setProgress(0);
@@ -186,6 +181,19 @@ public class ExpressExerciseActivity extends AppCompatActivity implements TextTo
         countDownTimer.cancel();
         startPauseButton.setText("START");
         timerRunning = false;
+    }
+
+    public void updateTimer() {
+        int minutes = (int) timeLeftinMills / 60000;
+        int seconds = (int) timeLeftinMills % 60000 / 1000;
+
+        String timeLeftText;
+        timeLeftText = "" + minutes;
+        timeLeftText += ":";
+        if (seconds < 10) timeLeftText += "0";
+        timeLeftText += seconds;
+
+        timer.setText(timeLeftText);
     }
 
     @Override
