+++ app/src/main/java/com/example/alafitness/LoginActivity.java
@@ -9,6 +9,10 @@ import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+/**
+ * Class for logging into the application, child class that contains inherited methods.
+ * Calls methods from the DBHelper class.
+ */
 public class LoginActivity extends AppCompatActivity {
 
     EditText username, password;
@@ -54,7 +58,7 @@ public class LoginActivity extends AppCompatActivity {
         buttonSignup.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent2 = new Intent (LoginActivity.this, MainActivity.class);
+                Intent intent2 = new Intent (LoginActivity.this, SignupActivity.class);
                 startActivity(intent2);
             }
         });
